All objects split
All objects split successful. Cimg is failing for .png file. Should find alternative.
