+++ lightObject.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 class lightObject
 {
     private:
