+++ normalObject.hpp
@@ -7,9 +7,10 @@ using namespace Eigen;
 
 class normalObject
 {
-    public:
+    private:
         RowVector3d *surfaceNormal; //Storing a set of normals, where normal at index i is the normal for the face i
         int number_of_normals;
+    public:
         void set_size_of_normals(int faceCounter);
 
         void generate_normals(displayObject object);
