+++ sharedVertexObject.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 #include <vector>
 #include <stdlib.h>
 #include <iostream>
