+++ sphereObject.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 #include "Eigen/Dense"
 #include "sphereObject.hpp"
